requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
chex==0.1.90
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.14
decorator==5.2.1
dm_pix==0.4.4
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.7
-e git+https://github.com/matthias-wright/flaxmodels.git@09bc77215032375d124e3f1eba828dd89c80a850#egg=flaxmodels
fonttools==4.58.5
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.34.3
humanize==4.12.3
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.1
namex==0.1.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.26.4
numpy-groupies==0.11.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.5
optree==0.16.0
orbax-checkpoint==0.5.16
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
ribs==0.8.1
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.32.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.19.0
tensorstore==0.1.76
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
treescope==0.1.9
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vendi-score==0.0.3
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
